refactor: spread scenes into multiple files
